qcacld-3.0: Cache key install flag on key install confirmation event
According to the current design, while executing the add key
command, the driver holds the northbound context (wpa_supplicant)
until it receives confirmation of key installation. Once the
confirmation is received, it posts the set key response
(WMA_SET_STAKEY_RSP) to the scheduler thread and caches key
install flag in STA DS node.

But sometimes, the driver receives the key installation event and
the action frame (such as ADDBA request frame) from southbound in
rapid succession.
However, the ADDBA frame processing is scheduled before the
WMA_SET_STAKEY_RSP is posted to the scheduler.
As a result, the driver processes the ADDBA request and sends the
ADDBA response to the firmware before releasing the supplicant
context and without setting the key-installed flag.
Consequently, the WEP bit is not set in the frame header, which may
lead to unencrypted frame being transmitted over the air.

So, to fix this, cache key installed flag in the peer object upon
receiving the key installation confirmation event.

Change-Id: Ic522c1db88a5950bdf7bd4cb5225cb7ce8da41a9
CRs-Fixed: 4230446