cnss2: Fix build errors with upstream MHI by isolating downstream-only fields
When building the cnss2 driver with the upstream kernel,
compilation errors occur due to the use of certain
mhi_controller structure members that are only defined
in the downstream MHI driver.

To resolve this, move the initialization of these
downstream-only members into a separate helper function,
and guard it with #if IS_ENABLED(CONFIG_MHI_BUS_MISC) to
ensure compatibility with both upstream and downstream MHI
implementations.

Change-Id: I1caa1e530f8e321c29cb8809d360e6a517cc3553
CRs-Fixed: 4230182