qcacld-3.0: Indicate Radar only when CAC is required
During CAC, if a radar event triggers a punctured channel or causes
a switch to a new DFS channel—either entering or exiting CAC—the CAC
process must be restarted. However, once SAP is UP, if a radar event
updates the punctured channel, CAC should be avoided.

Currently, radar indications are always forwarded to hostapd as soon
as they are received from the target. This behavior can cause issues:
if hostapd receives a CAC END event after a radar indication has
already been received, it will not transition to the UP state.

To prevent this, the driver should avoid sending radar indications
to hostapd unless the radar event actually triggers a channel restart
or a puncture update during CAC.

Change-Id: Ibb852df70ca10512c4e924e43968783887d97d9d
CRs-Fixed: 4253357