cnss2: fix modpost error by dropping cma_get_name
Previously, the module used cma_get_name() to log the
name of the CMA area. This worked on Android Linux kernels
where the symbol is exported.

However, in upstream Linux kernels, cma_get_name() is not
exported, leading to a modpost error during out-of-tree
module build.

To fix this, the call to cma_get_name() has been removed and
replaced with a simple validity check on the CMA area.

Change-Id: I5db4fca05cbad001a8d0853860ae68bc41a3a920
CRs-Fixed: 4263707