cnss2: Mark __cnss_pci_add_fw_prefix_name as static
This functions __cnss_pci_add_fw_prefix_name is only
used within pci.c, but it was not declared as static,
resulting in a -Wmissing-prototypes error during
compilation with Clang.

This change marks this functions as static to restrict its
scope to the translation unit and resolve the build issue.

Change-Id: I2a62959758ec430f5d0323e49ee25081a2b0dad2